refactor: inline custom scalars logic
