+++ src/extensions/SchemaErrors/injectTypenameOnRootResultFields.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from 'vitest'
 import { normalizeOrThrow } from '../../documentBuilder/Select/document.js'
 import { SelectionSetGraphqlMapper } from '../../documentBuilder/SelectGraphQLMapper/__.js'
-import { graffleMappedResultToRequest } from '../../layers/5_request/core.js'
+import { graffleMappedResultToRequest } from '../../requestPipeline/core.js'
 import { injectTypenameOnRootResultFields } from './injectTypenameOnRootResultFields.js'
 import { GraffleSchemaErrors } from './tests/fixture/graffle/__.js'
