+++ src/layers/6_client/Settings/Config.ts
@@ -1,8 +1,8 @@
 import type { GraphQLSchema } from 'graphql'
 import type { RequireProperties } from '../../../lib/prelude.js'
+import type { TransportHttp, TransportMemory } from '../../../requestPipeline/Transport.js'
 import type { GlobalRegistry } from '../../../types/GlobalRegistry/GlobalRegistry.js'
 import type { SchemaDrivenDataMap } from '../../../types/SchemaDrivenDataMap/__.js'
-import type { TransportHttp, TransportMemory } from '../../5_request/Transport.js'
 import type { Extension } from '../extension/extension.js'
 import type { TransportHttpInput } from '../transportHttp/request.js'
 import type { InputStatic } from './Input.js'
