+++ src/layers/6_client/client.transport-http.test.ts
@@ -4,7 +4,7 @@ import { serveSchema } from '../../../tests/_/lib/serveSchema.js'
 import { Graffle as Pokemon } from '../../../tests/_/schemas/pokemon/graffle/__.js'
 import { Graffle } from '../../entrypoints/main.js'
 import { ACCEPT_REC, CONTENT_TYPE_REC } from '../../lib/grafaid/http/http.js'
-import { Transport } from '../5_request/Transport.js'
+import { Transport } from '../../requestPipeline/Transport.js'
 
 const schema = new URL(`https://foo.io/api/graphql`)
 
@@ -32,7 +32,7 @@ test(`anyware hooks are typed to http transport`, () => {
 })
 
 import { schema as schemaPokemon } from '../../../tests/_/schemas/pokemon/schema.js'
-import type { CoreExchangeGetRequest, CoreExchangePostRequest } from '../5_request/hooks.js'
+import type { CoreExchangeGetRequest, CoreExchangePostRequest } from '../../requestPipeline/hooks.js'
 
 test(`when envelope is used then response property is present even if relying on schema url default`, async () => {
   const service = await serveSchema({ schema: schemaPokemon })
