+++ src/layers/6_client/extension/extension.ts
@@ -4,8 +4,8 @@ import type { FnProperty } from '../../../lib/fluent/Fluent.js'
 import type { AssertConstraint } from '../../../lib/prelude.js'
 import type { TypeFunction } from '../../../lib/type-function/__.js'
 import type { Fn } from '../../../lib/type-function/TypeFunction.js'
+import type { RequestCore } from '../../../requestPipeline/__.js'
 import type { GlobalRegistry } from '../../../types/GlobalRegistry/GlobalRegistry.js'
-import type { RequestCore } from '../../5_request/__.js'
 import type { Client } from '../client.js'
 import type { ClientContext } from '../fluent.js'
 import type { GraffleExecutionResultEnvelope } from '../handleOutput.js'
