+++ src/layers/6_client/handleOutput.ts
@@ -10,8 +10,8 @@ import {
   type GetOrNever,
   type Values,
 } from '../../lib/prelude.js'
+import type { TransportHttp } from '../../requestPipeline/Transport.js'
 import type { GlobalRegistry } from '../../types/GlobalRegistry/GlobalRegistry.js'
-import type { TransportHttp } from '../5_request/Transport.js'
 import type { RunTypeHookOnRequestResult } from './extension/extension.js'
 import type { ClientContext } from './fluent.js'
 import {
