+++ src/layers/6_client/requestMethods/requestMethods.ts
@@ -4,9 +4,9 @@ import { SelectionSetGraphqlMapper } from '../../../documentBuilder/SelectGraphQ
 import type { TypeFunction } from '../../../entrypoints/utilities-for-generated.js'
 import type { Fluent } from '../../../lib/fluent/__.js'
 import { isSymbol } from '../../../lib/prelude.js'
+import { RequestCore } from '../../../requestPipeline/__.js'
+import { graffleMappedResultToRequest } from '../../../requestPipeline/core.js'
 import type { GlobalRegistry } from '../../../types/GlobalRegistry/GlobalRegistry.js'
-import { RequestCore } from '../../5_request/__.js'
-import { graffleMappedResultToRequest } from '../../5_request/core.js'
 import { type ClientContext, defineTerminus } from '../fluent.js'
 import { handleOutput } from '../handleOutput.js'
 import type { Config } from '../Settings/Config.js'
