+++ src/requestPipeline/hooks.ts
@@ -1,10 +1,10 @@
 import type { FormattedExecutionResult, GraphQLSchema } from 'graphql'
-import type { Grafaid } from '../../lib/grafaid/__.js'
-import type { getRequestEncodeSearchParameters, postRequestEncodeBody } from '../../lib/grafaid/http/http.js'
-import type { httpMethodGet, httpMethodPost } from '../../lib/http.js'
-import type { ClientContext } from '../6_client/fluent.js'
-import type { Config } from '../6_client/Settings/Config.js'
-import type { MethodModeGetReads, MethodModePost } from '../6_client/transportHttp/request.js'
+import type { ClientContext } from '../layers/6_client/fluent.js'
+import type { Config } from '../layers/6_client/Settings/Config.js'
+import type { MethodModeGetReads, MethodModePost } from '../layers/6_client/transportHttp/request.js'
+import type { Grafaid } from '../lib/grafaid/__.js'
+import type { getRequestEncodeSearchParameters, postRequestEncodeBody } from '../lib/grafaid/http/http.js'
+import type { httpMethodGet, httpMethodPost } from '../lib/http.js'
 import type { TransportHttp, TransportMemory } from './Transport.js'
 
 interface HookInputBase {
